requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
django==3.2.10
django-admin-interface==0.18.4
django-colorfield==0.6.3
django-crispy-forms==1.13.0
django-environ==0.8.1
django-filter==21.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.0.0
psycopg2==2.9.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
whitenoise==5.3.0
